refactor: update imports for filter and utils to correct paths and enhance type definitions in filter.ts and utils.ts
